Hackthon/Backend/requirements.txt
google-adk
google-cloud-storage
google-cloud-vision
pymupdf
fastapi
google-cloud-speech
pydub
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
uvicorn
uvicorn[standard]
google-cloud-texttospeech
google-cloud-speech
python-socketio[asyncio_server]==5.12.1
python-multipart==0.0.6
requests==2.32.0
deprecated==1.2.14